refactor: complete SDL3 audio subsystem overhaul
- Fix 48000 Hz crackling: dynamic chunk sizing (~100ms) instead of 4096 bytes
- Voice pool for SFX: up to 8 simultaneous voices per sound id
- IAudioDecoder + DecoderRegistry: decoder-agnostic file loading
- AtomWavDecoderBackend uses SDL_LoadWAV internally (UTF-8 path fix)
- Log system: SDL.Backend.* channels, default level INFO, throttled push logs
- CMakeLists: renumbered sections, engine_log linked to audio libraries
- Code style: fix include order, Doxygen headers, SFXEntry member naming
- Fix ResizeEvent narrowing warning

+++ Engine/Render/SDL3RenderWindow.cpp
@@ -57,7 +57,8 @@ static auto TranslateEvent(const SDL_Event& ev) -> IEvent {
 
     case SDL_EVENT_WINDOW_RESIZED:
         result.type = EventType::Resized;
-        result.data = ResizeEvent{ev.window.data1, ev.window.data2};
+        result.data = ResizeEvent{static_cast<uint32_t>(ev.window.data1),
+                                   static_cast<uint32_t>(ev.window.data2)};
         break;
 
     default:
